+++ src/Products/PloneMeeting/tests/testViews.py
@@ -38,6 +38,7 @@ from zope.intid.interfaces import IIntIds
 
 import transaction
 
+
 SAMPLE_ERROR_MESSAGE = u'This is the error message!'