+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 4.1.12 (unreleased)
 -------------------
 
-- Nothing changed yet.
-
+- Adapted code to redefine the 'IGenerablePODTemplates' adapter for context and dashboard now that 'get_all_pod_templates' and 'get_generable_templates'
+  were moved from 'DocumentGeneratorLinksViewlet' to 'GenerablePODTemplatesAdapter' in 'collective.documentgenerator'
 
 4.1.11 (2019-11-19)
 -------------------